refactor: move NavigationTree to its own folder
